refactor: form-fields 구조 개선
- CurriculumFields와 PricingFields를 각 스텝에 직접 구현
- 도메인 특화 컴포넌트 분리로 일관성 향상
- 중복 필드 표시 문제 해결
- 공통 재사용 컴포넌트만 form-fields에 유지 (DatePickerFormField, DynamicListField)
